+++ docs/changelog.md
@@ -14,9 +14,11 @@
 
 ## v1.0-beta7 (2021-06-xx)
 
+- New overloaded procedure `add_param` to collect parameters before the call one
+  of the log methods error, warn, info, log, debug and trace
+- New procedure `printf`
 - Rename table column `CONSOLE_LOGS.LOG_SYSTIME` to `CONSOLE_LOGS.LOG_TIME` and
   change the data type from `timstamp` to `timestamp with local time zone`
-- New procedure `printf`
 - Remove prefix `get_` from all helper functions after reading [Stevens article
   about naming
   conventions](https://www.insum.ca/feuertip-11-what-makes-an-effective-naming-convention/)